fix: drive AF input meter from browser mic level
The radio's MIC/MICPEAK meters read silence for the browser's network
TX-audio path (verified on a FLEX-8000: MICPEAK ~-100 dBFS during active
TX), so the "AF Input Level" meter never reflected the transmitted mic.
The fill was additionally bound to the "MIC" meter, which never streams
(stuck at 0 -> full bar).

Extract the Web Audio level analysis out of AudioLevelMeter into a
reusable createStreamLevel() hook, and use it to drive the AF Input Level
meter from remoteAudioTxStream() — the microphone stream the browser
actually sends to the radio.
